LayoutTests/platform/gtk/media/W3C/video/canPlayType/canPlayType_two_implies_one_1-expected.txt

video.canPlayType() - codecs parameter logic

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS "probably" is "probably"

TEST COMPLETE
spec reference
